AdsLib: OOI: manually merge xift/dev-ooi; add AdsLibOOI project
copying headers from the libs into example/ was frustrating, error prone
and provoked duplicates:
-> use include path in example.vcxproj
-> move ooi code into a second library 'AdsLibOOI'

Include path for example project is '../' so we can use
it easier to see where the header comes from.

AdsLibOOI is meant to be a second library layer on top of AdsLib. The
idea is to have just and object oriented abstraction layer on top of
the existing libary, with the benefit, that we can use TcAdsDll on
TwinCAT systems for the lower layer!