AdsLib: harden AdsAddRoute()/AdsDelRoute()
To simplify the route and port handling for OOI, AdsRoute should hide port
allocation and route configuration automatically. This can be achieved by
reference counting the AmsNetId<->IpV4 pairs and deleting an AmsConnection
only, when all references have been released again.

There are 4 situations which have to be taken care of:
1. add route for a new netId and a new ip
 -> normal case create a new AmsConnection and mapping
2. add route for a new netId with an ip for which an AmsConnection already exists
 -> increment AmsConnection refCount and add a new mapping
3. add route for a known netId with a new ip
 -> this is not allowed as it would conflict with a previous route
4. add route which already exists
 -> increment refCount