AdsLib: directly use user supplied buffer for AoEResponses
commit #aa2ba7bfdba0c8576de4c6e09c06c5c0d19b3d5f was a very bad idea.
It allocates receive buffers dynamically depending on a length field
provided by an incoming AoE response ...

Fix this by using the user supplied buffer directly to receive the
incoming AmsResponse. This approach has the advantage that AdsLib
doesn't need to allocate any receive buffer. And additionally we can
save some cpu cycles by reading directly into the user buffer instead
of reading into a temporary buffer and later copy it again to the user.
The downside of this approach is, we now have to wait to respond to the
caller until no AdsLib thread is using the buffer anymore.
Identified scenarios:
1. response arrives and is processed within timespan
2. response arrives within timespan, but processing it exceeds timeout
3. response arrives to late or never

The first case is the good case. The processing would look like this:
A. An AmsResponse is reserved for the current AmsRequest and marked
   with an unique (32bit) invokeId.
B. The AmsRequest is send to the wire.
C. The calling thread is suspended until a response is received for the
   selected invokeId or a timeout occurs.
D. AmsConnection will receive a response with correct invokeId and atomically
   locks the user supplied buffer by consuming (reset to zero) the invokeId.
E. AmsConnecton receives all of the response directly into the locked user
   buffer.
F. The calling thread is notified and woken up
G. AmsResponse is released by atomically resetting the referenced AmsRequest

In the second case processing will be different as from step E. Because at
that time the calling thread will be woken up because of a timeout. But
instead of just releasing the AmsResponse, now the thread has to wait
for AmsConnection to completed receive, because the user supplied buffer
is already locked. The good thing is, the AmsConnection is using the same
deadline for socket.receive as the calling thread used to sleep. So the
additional delay should be small. The calling thread will now wait until
it is notified by AmsConnection and then cleanup and return as in the first
case.

The third case is easier. The calling thread is woken up as in the second
case. But this time the buffer isn't locked by AmsConnection but by the
woken up thread himself. So we can safetly continue as in the other cases.

For completness there is a fourth case in which the response is processed by
AmsConnection, even before the calling thread was suspended. That case
behaves exactly as case 1 (but was tricky during development;-).

Code changes:

- Deduct AmsResponse header type from responses cmdId
- Remove AmsResponse member 'frame' which was used as temporary receive buffer
- AmsResponse::Wait() is now waiting until a precalculated deadline and
  returns the errorCode of the current AmsRequest/Response.
- AmsConnection::Release(AmsResponse*) is refactored as AmsResponse::Release()
- AmsConnection::Receive() with deadline was added to minimize delays
- AmsConnection::ReceiveFrame() was refactored to avoid delays by using deadlines
- wait_until deadlines instead wait_for timeouts to minimize delays
- AmsConnection::AdsRequest is much simpler, now.

 # Last command done (1 command done):