Revert "Makefile: Use make env-variable for os detection"
It was a bad idea to rename OS_NAME to OS. This breaks all existing
build scripts, where OS_NAME was set manually. Even worse the patch
doesn't seem to fix the original problem so it makes no sense to keep
the patch.
If you want to have auto detection but uname is not available I suggest
make it availabe on Windows with something like:
echo echo win32 > uname.bat

This is related to #84
This reverts commit 646f755b53aa3a77cfb62845a18690631c0bf636.