OOI: Object-Oriented interface
A long time ago, Michael suggested to improve AdsLib with an object oriented
interface. That interface is meant to be a second library layer on top of
AdsLib. This way we can reuse this new OOI on top of TcAdsDll of TwinCAT
system, too!

This commit finally merges some classes to our master branch:

- AdsHandle, wrapper to help cleaning up ADS handles
- AdsVariable, helper to read and write simple ADS variables
- AdsNotification, helper to work with notifications
- AdsDevice, main object to communicate with some ADS remote

Suggested-by: Michael Purwins <mpurwins@gmail.com>