Sockets: throw an exception, when TCP connect() failes.
Currently an error during TCP connect will not be propagated to the the
application code. It is hidden within AmsConnection::ownIp (==0).
To improve this we now throw a std::runtime_error, which will prevent
AmsRouter::AddRoute() to emplace a broken AmsConnection into its list
of available AmsConnections.
This was the problem with GH-113. On the first try a new AmsConnection
would be constructed and added to the list. On the next retry AddRoute()
would look up the target AmsNetId again, find it, but with a broken
object, which could only be recovered by manually deleting the route.
It's much better to not add broken AmsConnections to the list in the
first place.