AdsNotification: prepare to use with TwinCAT router
In AdsLib we use pointers to constant objects in the notification callback
function. TcAdsDll uses non-const. To somehow change this in a backwards
compatible manner, we add new typedefs for both variants and implement
AdsNotification constructors for both variants, which cast to the generic
PAdsNotificationFuncEx, which is defined depending on which ADS implementation
was choosen.

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>