AdsLib: prepare abstraction layer to use TwinCAT system headers
We want to be able to use our object oriented interface on TwinCAT systems, too.
In order to use TcAdsAPI.h and TcAdsDef.h on these systems, we have to add an
indirection for our existing headers AdsLib.h and AdsDef.h.
Here we move them into AdsLib/standalone and only keep additional generic extensions
such as operator overloads for AmsAddr and AmsNetId in general file.

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>