AdsLib: repair async notifications
commit 8e4e8bb9d656097e7bbe3d19d18cffeb9e0ab37e was pretty bad. We
intended to just run NotificationDispatcher::Run() asynchronously on a
different thread. But because we didn't store the std::future returned
by std::async(), the destructor of that temporary std::future
immediately gets called. So effectively we start a new thread, give him
a job to do and block until he is done.

So our first solution wasn't that bad. We had one thread per
notification and one ringbuffer per notification, so we are just fine.
The only bad thing is that we have to implement a semaphore. This will
change as soon as we use C++20.

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>