ci: build with meson on Windows 10
Having both Makefiles and Visual Studio Solutions was always bad to maintain.
Using CMake to have a simple build file for both usecases seemed promising, but
was full of conditionals, because CMake is unable to abstract things like
threading library or warning levels. Meson does these things, well and it was
pretty easy to write one meson.build script for Windows and all of our other
platforms. This commit just introduces the meson.build for a native Windows
build.

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>