ci: use a container as ads-server
With GitLab we can provide service containers, which are automatically added
to the virtual network of our test_runner container. Lets take one with a
test version of ads-server, so we can run multiple tests in parallel!
This should speed up our pipeline significantly and will allow us to test more
flavours of AdsLib. Like TcAdsLib or different compilers.

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>