Tests: make version check less strict
When we test AdsSyncReadDeviceInfoReqEx() we are very strict at checking the
correct TwinCAT version of our ads-server. This proofed to be very annoying,
as it requires us to update the test for each upgrade we do to ads-server.
In the future we might want to test against different ads-server instances
with different versions, so lets only check it the build is at least 1711.

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>