AdsLib: move routing functions into common header
AdsAddRoute(), AdsDelRoute() and AdsSetLocalAddress() are functions specific to
AdsLib and not part of TcAdsDll.dll. However, we need dummy implementations for
the TwinCAT Libary, so we can hove the redundant definitions into our common
AdsLib.h. We soon want to introduce a function to add routes to remote targets,
too. So we use this move to rename the functions and put them into bhf::ads
namespace. To be backwards compatible we define the old names to point to the
new functions in the namespace.

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>