ci: always add a new ADS route in our tests
Previously, we had to use predefinded routes for our FreeBSD and macOS test
runner machines, because pyads was not available on them (or we have been to
lazy to setup). However, since AdsTool supports adding routes over UDP, we can
use it on all supported AdsLib platforms, so lets always add the route during
testing. FreeBSD is VM on the same host as the docker container running our
ADS dummy PLC. There we need NAT, so the IP in our route is different.

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>