AdsTool: add netid command
In ADS it's pretty common, that we only know the IP or hostname of our TwinCAT
system. But for ADS communication we need to know the AmsNetId. So we add a new
command to our AdsTool to ask a host for its AmsNetId or to be more precise ask
the main TwinCAT router on that system for its AmsNetid.

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>