ci: replace CX2000 WinCE ads-server with container on lbs1
It makes no more sense to have dedicated hardware for our AdsLib tests. Our
Linux tests already use a companion container so we use something similar for
FreeBSD and macOS tests. As it's still not trivial to implement companion
containers just as easy like for docker runners we dedicated a container on
lbs1 to act as the new ads-server for physical and VM test runners.
To have fresh containers, we restart them by connecting to port 48897, which
is forwarded into the container and will kill it.
We use the opportunity to switch our FreeBSD VM, used for testing. Now, we
use the same we used for building, which is based on FreeBSD 13. As it runs
NATed we have to add our route at ads-server with that NAT IP!

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>