ci: install vim instead of vim-console for FreeBSD
FreeBSD project is completely broken. We use a vanilla FreeBSD 13 VM for
testing. Unfortunately, we need xxd which is part of vim. vim-console is a
small package containing vim and xxd so we choose that, after vim-tiny was
broken in FreeBSD previously. Well, today vim-console got broken, too.
https://www.freshports.org/editors/vim-console shows it is available
from quarterly, but not for latest. WTF!!!!!

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>