AdsTool: show usage on stderr instead of stdout
In general we reserve stdout for data and stderr for error messages. Somehow
our usage() ended up on stdout. There should be only one case in, which this
is desired behavior: When we are explicitly asked for the usage(). So we use
the opportunity to cleanup stderr and add an explicit --help.

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>