AmsRouter: deprecate AddRoute(AmsNetId, const IpV4&)
Hardcoding IpV4 into a public API was pretty dumb. Downstream projects
should upgrade to the new hostname based API as soon as possible.

For this we now deprecated the old API and upgrade our own testcases.

[[deprecated]] was introduced with c++14. Since the README says "build this
library with a recent compiler", no one should be overwhelmed, if we
raise our minimum requirements a bit to an 8 year old standard.

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>