TcpSocket: implement bool IsConnectedTo(addrinfo*)
At the moment our whole AdsLib implementation depends on the 32 bit
representation of an IPv4 address to lookup AMS routes for AmsNetId
targets. In order to change this, we need a mechanism to find out if for
a given hostname, we already have a running TCP connection.
With this commit we add the very basic infrastructure for this.
IsConnectedTo() asks for a preallocated list of "struct addrinfo*" like
getaddrinfo(3) would return. If you pass in such a list, IsConnectedTo()
will match that list against the sockaddr of its internal TCP socket. On
first match IsConnectedTo() will return true. If no match is found it
returns false.

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>