AmsRouter: add AddRoute(AmsNetId, const std::string&)
Hardcoding IpV4 into a public API seems pretty "limiting". In order to
prepare AdsLib for IPv6 usage, we first have to get rid of this madness.
For compatibility we keep the old AddRoute(AmsNetId, const IpV4&), but
we will pretty soon mark it deprecated to make sure we drop it fast
from our callsites.

The main problem is our entire AmsConnection lookup depended on these 32
bit IpV4 addresses. We even used that address as a key in our
connections mapping. However, reality shows many TCP connections are
pretty rare. Additionally, lookups in our connections map happen only,
when we are adding or removing AMS routes. So changing the datastructure
for "connections", shouldn't have a big impact. Since, these lookups
look so unimportant performance wise, we even go for an unordered_set,
because it doesn't require == and < operators to be added for
AmsConnection. We could easily improve that, if someone really hits
issues with this.
So we drop IPv4 to be our primary connection identifier, what do we use
instead? Well, we still use it, but in a more abstract form. Our
TcpSocket already gained a new function to check if it is connected to a
list of possible addresses (struct addrinfo). So we know compare the
sockets sockaddr_in with the addresses from the addrinfo list. As of now
this is still IPv4 only, but IPv6 is now in reach and will follow soon.

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>