AdsLib: implement IPv6 support
This should be the last step to finally support IPv6. The last remaining
issue, where our sockets, which depended on an IpV4 parameter to find
their target. We replaced that parameter with a list of
"struct addrinfo*", which in case of a TcpSocket we iterate and try
every entry until we find one, which accepts our connection.

AmsConnection lost it's destIp member, which was previously used to
lookup the correct AmsConnection for a new route. This was made obsolete
by Socket::IsConnectedTo() and the last user of this member was our
unittest. If some downstream project depends on this member, it can be
easily reintegrated, by reading from m_Socket.m_SockAddress.

The most important thing about IPv6 is the new implicit local AmsNetId.
If not set explicitly with bhf::ads::SetLocalAddress(), the AmsNetId is
by default set to "<ip>.1.1" for the first AmsConnection. Or for an IPv6
connection always set to "255.255.255.255.1.1". So you might make want
to make it a habit to bhf::ads::SetLocalAddress(), before you add any
route.

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>