AdsLib: support variable ADS ports
TwinCAT ADS router always listens on TCP port 48898 and UDP port 48899
(ads-s at TCP/8016 not supported by AdsLib). But in todays factories it
is not unusual to multiple TwinCAT hosts (virtual or physical) behind a
firewall or gateway. To allow port forwarding for different TwinCAT
hosts behind the same gateway, we need ouur AdsLib to be able to connect
to different TCP ports than 48898.
To support this with minimal changes to existing interfaces, we adapt a
host specifier format similar to RFC2732, e.g.:
localhost:12345
127.0.0.1:12345
[::FFFF:127.0.0.1]:12345
[2605:2700:0:3::4713:93e3]:12345

If no port is found in the host string we fall back to TCP/48898 and UDP/48899

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>