ci: test variable TCP port
We claim we can connect to TwinCAT routers listening on different ports
than 48898, lets proof it by creating a socat port forwarding and
connecting to the local socat port.
At the moment socat is not easily available for TC/BSD as soon as that
changes, we will expand our test to TC/BSD, too. Windows might be more
difficult, we care about that later, when we care about Windows again.

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>