AdsDef: deprecate enum nSystemServiceOpenFile
We introduced it to support file transfers over ADS. Well, recently we
discovered it was bad practice to mix enums in a bitmask and we can
replace SYSTEMSERVICE_OPENGENERIC with FOPEN::READ.
Since it was declared in a public header, we don't remove it
immediately. The risk someone else than AdsTool is using it should be
prettly low, but keeping it a while shouldn't hurt to much.

[[deprecated]] is only supported in c++17 so until we raise our
compiler requirements, we have to make this a soft deprecation.

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>