AdsTool: implement --retry
ADS/AMS is not a reliable transport layer such as TCP. So it is very
common, that we have to repeat ADS requests. The adstool is heavily used
in CI pipelines, where we often need determinism. So we have workarounds
in place, which retry specific commands and only give up after to many
tries.
To get rid of all these custom retry wrappers, it seems reasonable to
implement this retry behaviour directly in adstool.

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>