ci: use new BHF_CI_NAT_IP
Our test infrastructure has changed. The adslib tests don't run on lbs1
anymore, but on pc-3d476b. This is just a hotfix to repair the broken ci
pipeline. A better approach would be to use a qemu docker container with
luemuctl and TC/BSD inside, which run against a service container, just
like we do it for our Linux containers.
But as we have an important feature (--retry) pending, we add this
hotfix, now and fix the ci later.

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>