adstool: fix exit codes for license commands
c and shell boolean values are inverted, i.e. false means 0 in c which
then means success for the shell when returned as exit code.
In case of an error, the license functions return false, with applies
to the c/c++ conventions, but makes the shell think that the command
succeeded. Changing this to 'if (error) return true' would return the
correct exit code, but would be confusing to read.
Therefore the interface of these functions is changed to int to
workaround any bool-conversions.

Signed-off-by: Julian Heuking <J.Heuking@beckhoff.com>