ci: prepare a test-stage test for AdsLib
We want to run AdsLib tests on a TC/BSD target. For this we can easily
reuse our test-stage tool, which creates a QEMU VM within our generic
test-runner container and installs TC/BSD into this VM.
With this we cann then simply copy our build artifacts and test scripts
into the VM and run our tests there just like we already do within our
Linux containers.
One difference is routing, but we already have a mechanisim for NAT
routing, which we can reuse here to allow the QEMU VM to use the NAT of
the test-runner container to communicat ADS to the docker service
container, which contains our test TwinCAT endpoint.

NOTE: The vim package on TC/BSD provides xxd, which is required by the
tests.

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>