Sockets: change m_DestAddrLen type to match socklen_t
Previously, size_t was no problem, but when we add MSVC to our compiler
mix we get many conversion warnings. So lets align our types with types
from struct addrinfo. Well, even better than "socklen_t" would be
"decltype(addrinfo::ai_addrlen)", but of course MSVC prevents us from
using that.
The reason is a bit complicate. MSVC uses size_t for ai_addrlen, but
wants to get int values passed to functions like connect() or sendto().
So we could either cast from size_t to int whenever we call connect()
or sendto(), or we cast only when we read from ai_addrlen and store the
value into m_DestAddrLen.
We choose the second, because it is simpler to implement.

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>