AdsTool: fix binary output on stdout on Windows
Some popular C/C++ libraries on Windows automatically replace "\n" with
"\r\n", when writing to stdout

  # Let's have a look at our test string in hex:
  $ printf 'aaa\naaa' | xxd
  00000000: 6161 610a 6161 61                        aaa.aaa

  # Now, send it to some random TwinCAT device:
  $ printf 'aaa\naaa' | ./tcadstool.exe 5.75.184.60.1.1 file write 'text.txt'

  # And read it back with tcadstool.exe running on a Windows machine:
  $ ./tcadstool.exe 5.75.184.60.1.1 file read 'text.txt' | xxd
  00000000: 6161 610d 0a61 6161                      aaa..aaa

  # Spot the difference:
  # Send: 00000000: 6161 610a 6161 61     aaa.aaa
  # Recv: 00000000: 6161 610d 0a61 6161   aaa..aaa

Well, thanks but no thanks this is an antifeature, which we have to disable
at least, when we write binary data.

If we would disable this behaviour globally it might hurt "text" output
for other commands, when Windows tools wouldn't print a line break for
'\n' without '\r'.

So we don't set stdout to binary mode globally, but only when we expect
binary data. Let's see how far we get with this.

Reported-by: Jan Ole Hüser <J.Hueser@beckhoff.com>
Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>