ci: allow different platform level and systemids
Previously, we tried to match exact platform levels and systemids to
detect errors in platform level detection and systemid generation.
However, we have heterogenous systems in our CI so we should find a way
to handle different systemid in the same pipeline.
So for now we try to keep a list of systemid platformid pairs and match
against this list.

Suggested-by: Steffen Dirkwinkel <s.dirkwinkel@beckhoff.com>
Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>