SymbolAccess: implement FetchSymbolEntries()
It seems many users are interessted in a generic way to access PLC
variables. We already had a pull request, suggesting an implementation:
https://github.com/Beckhoff/ADS/pull/170

That implementation relied on opportunistic C macros for parsing of
networkdata, which by todays security standards is to risky and error
prone.

So here we implement the basic network data parsing. We introduce a new
structure bhf::ads::SymbolEntry, which contains all the information we
can gather from a TwinCAT PLC about its symbols. The function
SymbolAccess::FetchSymbolEntries() will return a map for all the
symbols, where the symbol name can be used to access the detailed symbol
information.

In the next steps we will implement some additional functions, which
make use of FetchSymbolEntries() to demonstrate how this map can be used
for generic PLC symbol access.

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>