Add compile time option CONFIG_DEFAULT_LOGLEVEL
With commit c1739cd6d6c551225f68cca2eb234eaba840b71f we added the
possibility to adjust the default loglevel at runtime, but it seems this
is not enough:
https://github.com/Beckhoff/ADS/pull/174

In the general usecase where AdsLib is integrated in a larger program it
might be a better option to configure the default loglevel at compile
time. This omit the call to Logger::logLevel at runtime, which might be
jus as annoying as WSAStartup()...

The original pull request disables dynamic behaviour completely, which
is not what we want.

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>