AdsTool: add state --compare command
Many times in CI we have to wait for the TcSystemService to accept ADS
commands. For this we already build many wrappers, with all their quirks
and problems.
This new command should make them obsolete, as we now have a call to the
adstool, which reads the current ADS state and compares it with one or
multiple values. If the read ADS state matches one of the provided
states, adstool will return 0, otherwise 1. Combined with --retry this
can be used to wait for the TcSystemService to become ready:
$ adstool 5.24.37.144.1.1 --retry=60 state --compare 5 15

Or other ADS devices like a TwinCAT3 PLC:
$ adstool 5.24.37.144.1.1:851 --retry=60 state --compare 5

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>

squash! AdsTool: add state --verify command

rename verify back to compare