tools: add test for startprocess
This seems overly convoluted for what it is supposed to test but it
should at least guarantee that our implementation is behaving as expected.
As the spawned process won't return it's exit code (since we will only
be able to know whether a process was created or not) we make sure the test
script actually ran by trying to execute another script which our previous
one creates.
We use /usr/bin/env to support FreeBSD and Linux targets.

Reviewed-by: Patrick Bruenn <p.bruenn@beckhoff.com>