Use .editorconfig to configure indent size & style
The file .editorconfig is recognized by many editors. We use it to make
it more probable that every file of the same programming language is in
the same format.

If we use tab characters for indentation for a language, we do not
specify the size of the tab. Everybody can have it's own tab size.