RegistryAccess: Check network answer when enumerating
When enumerating registry keys or values, check if the data from the
network is in the format we expect.

If we ask for registry keys, we don't want to get values, even if we
could parse them. And if we ask for registry values, we don't want to
get keys.

Signed-off-by: Jan Ole Hüser <J.Hueser@beckhoff.com>
Reviewed-by: Patrick Bruenn <p.bruenn@beckhoff.com>