RegistryAccess: Extend .reg file format for deletion of keys
If the key to be deleted cannot be found, we ignore the failure, because
of two reasons:

- regedit.exe also doesn't report an error in that case

- In either case (deletion succesfull / key not found) the result is the
  same: The key is not present afterwards.

But there are also reasons against ignoring the failure: Maybe the user
wanted to delete another key and made a typo, then they doesn't notice,
that they made an error.

The error code ADSERR_DEVICE_NOTFOUND is used to signal, that the key
couldn't be found. Maybe that will change, because the pull request for
the Runtime isn't merged yet.

If we have a regfile like this:

```
[-HKEY_CURRENT_USER\KeyToBeDeleted]
"ValueName"="Data"
```

Then we ignore the last line, because regedit.exe also ignores such
lines.

Signed-off-by: Jan Ole Hüser <J.Hueser@beckhoff.com>