fructose: Don't define function in namespace std
Some time ago we changed fructose to be compatible with C++14.  The
problem was, it used the function std::uncaught_exceptions (with an "s"
at the end) which was introduced in C++17. So we implemented
std::uncaught_exceptions in one of the fructose headers. But because the
functions was neither `static` nor `inline` the linker reported
"duplicate symbol", if the header was included in multiple translation
units: A violation of the One Definition Rule.

It seems like, the obvious solution is, to make the function inline or
static, like this:

```
  #ifndef __cpp_lib_uncaught_exceptions
  namespace std {
      static int uncaught_exceptions() noexcept {
          // ...
      }
  }
  #endif
```

But on FreeBSD we get this error:

  ../tools/fructose/test_root.h:64:16: error: static declaration of 'uncaught_exceptions' follows non-static declaration
      static int uncaught_exceptions() noexcept {
                 ^
  /usr/include/c++/v1/exception:134:63: note: previous declaration is here
  _LIBCPP_FUNC_VIS _LIBCPP_AVAILABILITY_UNCAUGHT_EXCEPTIONS int uncaught_exceptions() _NOEXCEPT;
                                                                ^

The standard library implementation on FreeBSD, does not define the
feature test macro, but declares the function, maybe because the
function is not defined.

The solution is to not define a function in the namespace `std`, but to
define a new function in the `fructose` namespace, which uses the
available standard functions and is `inline`.

Signed-off-by: Jan Ole Hüser <J.Hueser@beckhoff.com>
Reviewed-by: Patrick Bruenn <p.bruenn@beckhoff.com>