ci: Publish test-pack with AdsTest tool
... to test the deletion of registry keys via ADS.

The new AdsTest tool is rather a test of the system service
implementation so we keep it separate from our existing tests and put it
into .test-pack.

We add a test-pack job to put that folder into a test-pack and make it
available for downstream OS pipelines. We will trigger them in a follow
up commit.

Signed-off-by: Jan Ole Hüser <J.Hueser@beckhoff.com>