AmsRouter: Stop holding global lock during connect
Previously, the global router lock was held during TCP connection
attempt, this stalled all ADS communication, because the same lock is
used to access the AmsNetId to AmsConnection mapping.
We could implement a different lock for the different structures, which
would ease our problem, but still prevents multiple TCP connection
attempts in parallel.
Instead we implement an additional "waitqueue" for TCP connections,
which are being established. This has the advantage that we can now
speed up usecase, which establish multiple TCP connections to different
targets in parallel.
Our implementation might be problematic for users, which call AddRoute() for
different AmsNetIds, but the same TCP target. This should be avoided to
be done in parallel or only the last connection to a TCP target will
survive.

Fixes: https://github.com/Beckhoff/ADS/issues/208
Signed-off-by: Nicola Timeus <nicola.timeus@eurotech.com>
Reviewed-by: Patrick Bruenn <p.bruenn@beckhoff.com>