Debian: fix build dependencies
We use bdpg:master, which became a lot stricter on build dependencies
just recently, so lets fix ours.

Recently we discovered the Debian tool "mk-build-deps", which we can use
to automatically install all the build dependencies for bdpg.
"mk-build-deps" reads the debian/control file, creates a metapackage
with all the "Build-Depends" and installs the metapackage with apt.

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>