RegistryAccess::Write(): fix dword parsing/calculation
Our pointer trickery seemd to work fine previously, but when we upgraded
qemu-mips (Debian 11 -> 12) we always got:
  qemu: uncaught target signal 10 (Bus error) - core dumped

Still not sure why this worked before and is failing, now. However, we
want to go forward and move to Debian 12 so lets fix this with maybe
more inefficient but much simpler code.

Reported-by: Steffen Dirkwinkel <s.dirkwinkel@beckhoff.com>
Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>