ci: Null out variables in powershell environment
We stumbled up on a pretty weird behaviour, which we recognized from
another project:
TcBuild failed when we had a last commit containing this line in its
commit message:
  -smp [[cpus=]n][,maxcpus=maxcpus][,sockets=sockets][,dies=dies][,clusters=clusters][,cores=cores][,threads=threads]

https://git.beckhoff.dev/beckhoff/test_stage/-/commit/49b3684e4ae586f9ba0ef4cc13e34c61c5e74dc1

The cause of this is Enter-VsDevShell, which calls a batch script, which
then tries to interpret environment variables. '[[' is used by Microsoft
as a special marker in powershell. We stumbled up on this because by
accident our last commit message contained such a '[[' (like this commit
;-) and when the script runs in GitLab, GitLab will export the last
commit message into an environment variable (actually two).

One workaround would be to omit such characters in the last commit
message of our merge request. But this isn't stable and breaks
bisectability. Additionally it doesn't seem fair to force developers to
write workarounds in there commit message to bugfix Microsoft tools.

So we decided to unset the troubling environment variables explicitly
for the one job, in which we call Enter-VsDevShell. The error might
return when another variable contains magic characters, but at least we
then have a solution we can point to.

This time we found our commit message in the output of chocolatey. We
don't know which character it was this time, but applying a similar
patch helped again...

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>