ci: prepare for arm64 expansion
Soon, we want to build and test arm64, too. For this we want to use
native build and test environments just like for amd64.
Here we prepare the build and test jobs to be easily expanded for arm64.

Unfortunately, GitLab has no syntax to let a matrix test job
automatically depend on the corresponding matrix build job, so we will
have to copy & paste test-linux-gcc-amd64 for arm64.

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>