ci: build with bdpg container
In the past we used docker-mxe:40549, which is pretty old Debian
bullseye with mxe cross compiler preinstalled.

We still test mxe builds only because we can. Build of Windows binaries
is now tested natively in "build-win10" or cross compiled with clangcl.

To make the final drop of mxe support more easy, we migrate our build
jobs to use the most recent bdpg container. For this we have to install
meson manually.

For i386 builds we need to add multilib support. And for clang builds we
have to install clang manually.

Our gcc build became a bit more easy, because the new bdpg container has
our internal package server preconfigured, so we can install
libadscomm-dev directly.

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>