ci: use preinstalled meson for win10 builds
Windows is the worse platform to support. Yet again its failing to
build. This time chocolatey or the MSI installer complains it wants to
reboot after the upgrade of meson:
  Packages requiring reboot:
   - meson (exit code 3010)

So we simply drop the installation step for now. See you next time, when
we explicitly need a newer meson version!

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>