ci: switch to tcpkg runner
Our Windows agents are migrated to tcpkg. With tcpkg the TcAdsDll.dll is
not installed into C:\TwinCAT anymore, but instead into
C:\Program Files (x86)\Beckhoff\TwinCAT. Extra funny is the change in
the subdirectory. Previously, we had the static library in:
TcAdsDll\x64\lib not it is in TcAdsDll\lib\x64.

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>