ci: drop tools/80_ads_route.sh
Previously, we made use of BHF_CI_NAT_IP to overwrite the address used
in the ADS route. We need that special case when our test application is
running inside of a container or virtual machine and the host of that
container/vm is forwarding the ADS ports directly. In that case the
TwinCAT on the other side sees only the IP of the host, not of the
container or VM, so we have to use the hosts IP and not the container
IP.
During refactoring of our test stage tools, this resurfaced and we think
it is time to clean this up a bit and get rid of the environment
variable.
So we choose a straightforward approach and simply pass the hosts IP to
90_run_tests.sh in case we need the NAT special case.
If 90_run_tests.sh is called without additional parameter we just use
the containers own IP.

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>