DCAccess: Add Distributed Clocks Diagnosis
To measure the quality of the EtherCAT Distributed Clock deviation, an
option to diagnose the target/actual comparison of the EtherCAT frame
arrival times is added. The smaller the deviations, the better the
quality of the DC communication.
When a task is called, it calulates the expected time of its next call
based on the current time and its own cycle time. It compares the
expected time of its next call with the actual time at which this task
is called for the next cycle.

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>