ci: drop build-clangcl and tcpkg jobs
This job is broken[1][2] on hosts with many CPUs like kvm-build01.
Additionally, the build of the docker container, which is used for
build-clangcl is broken[3] too.
At the moment we don't have time to deal with problems in absolute
niches, so we are ending support for clangcl and tcpkg until a developer
finds time to deal with the issue.

[1] https://azdevops01.beckhoff.com/GSW01/TcOs/_workitems/edit/214122/
[2] https://git.beckhoff.dev/beckhoff/AdsLib/-/jobs/1594585
[3] https://git.beckhoff.dev/beckhoff/docker-tcpkg/-/jobs/1594435

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>