meson: use meson.project_source_root()
The old function meson.source_root() was deprecated since '0.56.0' so
here we go and follow the suggestion to use project_source_root()
instead.
Remember "example" is not a real subproject of AdsLib so we can't use
global_source_root() as you might would use in a real world project,
where AdsLib is the subproject of the example application.

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>