Drop CHANGES file, use debian/changelog instead
The package build for debian automatically took CHANGES and put it
into /usr/share/doc/adstool/changelog.gz.

CHANGES was abandond a long time ago. Instead we moved on to use
debian/changelog a while ago. So here we go and drop CHANGES to avoid
people confusing it with the actual changelog.

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>