ci: fix CMake and add a simple build test
We continue to support only meson. However, there seem to be enough
unfortunate users who are dependent on CMake. To avoid simple
regressions like forgetting cpp files for them in the future, we are now
adding a simple build test for CMake to the CI pipeline.

ATTENTION: If there are serious obstacles to using CMake in the future
we will revert this.

Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>