Doxygen: get some html output again
It looks like Doxygen no longer recognizes our pure C headers without us
adding a lot of Doxygen markers. Since we have neither the time nor the
desire to dance to the tune of a documentation tool, we try to squeeze
more output out of the tool with the crowbar. If this doesn't really add
value, we will simply drop the doxygen configuration.

Fixes: https://github.com/Beckhoff/ADS/issues/244
Signed-off-by: Patrick Bruenn <p.bruenn@beckhoff.com>